Add files via upload
(1) 1. Define two variables named as x and y , assign some integer value to x and y and
print (display) the value of x and y. 2. Perform following operations on x and y and display (print) the result with
suitable message. Multiplication, Division, Floor division, Modulus , Exponentiaion. 3. What is typecasting of data in python? Write example to convert a integer to
float, integer to string and string to integer . Check the data type of each .